requirements.txt
packageurl-python>=0.15.0
aiohttp>=3.9.0
semver>=3.0.0
